feat(iv): add fixed size intvec
