+++ succinct_neo/src/int_vec/dynamic.rs
@@ -1,33 +1,6 @@
-use crate::int_vec::{Dynamic, IntAccess, IntVec};
+use crate::int_vec::{num_required_blocks, Dynamic, IntVec, IntVector};
 
 impl IntVec<Dynamic> {
-    /// Gets the number of required blocks of the given type to contain the specified number of
-    /// elements of a given width.
-    ///
-    /// # Arguments
-    ///
-    /// * `num_elements` - The number of elements intended to be saved.
-    /// * `bit_width` - The bit width of each element.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use succinct_neo::int_vec::IntVec;
-    ///
-    /// // 32 * 10 makes 320 bits, requiring 5 * 64bit blocks.
-    /// assert_eq!(5, IntVec::num_required_blocks::<u64>(32, 10))
-    /// ```
-    #[inline]
-    pub fn num_required_blocks<T>(num_elements: usize, bit_width: usize) -> usize {
-        (num_elements as f64 * bit_width as f64 / (std::mem::size_of::<T>() as f64 * 8.0)).ceil()
-            as usize
-    }
-
-    #[inline]
-    fn recalculate_capacity(&mut self) {
-        self.capacity = self.data.capacity() * Self::block_width() / self.width;
-    }
-
     /// Creates an integer vector with a given bit width and a default capacity of 8.
     ///
     /// # Arguments
@@ -37,9 +10,9 @@ impl IntVec<Dynamic> {
     /// # Examples
     ///
     /// ```
-    /// use succinct_neo::int_vec::IntVec;
+    /// use succinct_neo::int_vec::{Dynamic, IntVec};
     ///
-    /// let v = IntVec::new(15);
+    /// let v = IntVec::<Dynamic>::new(15);
     ///
     /// // 8 integers of size 15 require 120 bits this in turn requiring 2 * 64 blocks (= 128 bits).
     /// // These can hold 8 integers exactly.
@@ -62,9 +35,9 @@ impl IntVec<Dynamic> {
     /// # Examples
     ///
     /// ```
-    /// use succinct_neo::int_vec::IntVec;
+    /// use succinct_neo::int_vec::{Dynamic, IntVec};
     ///
-    /// let v = IntVec::with_capacity(15, 20);
+    /// let v = IntVec::<Dynamic>::with_capacity(15, 20);
     ///
     /// // 20 integers of size 15 require 300 bits this in turn requiring 5 * 64 blocks (= 320 bits).
     /// // However, 320 bits can hold 21 integers of size 15
@@ -73,14 +46,14 @@ impl IntVec<Dynamic> {
     /// ```
     #[inline]
     pub fn with_capacity(width: usize, capacity: usize) -> Self {
-        let num_blocks = Self::num_required_blocks::<usize>(capacity, width);
+        let num_blocks = num_required_blocks::<usize>(capacity, width);
 
         let mut temp = Self {
             data: Vec::with_capacity(num_blocks),
             width,
             capacity: num_blocks * Self::block_width() / width,
             size: 0,
-            _marker: Default::default()
+            _marker: Default::default(),
         };
 
         temp.data.push(0);
@@ -100,9 +73,9 @@ impl IntVec<Dynamic> {
     /// # Examples
     ///
     /// ```
-    /// use succinct_neo::int_vec::IntVec;
+    /// use succinct_neo::int_vec::{Dynamic, IntVec};
     ///
-    /// let v = IntVec::new(15);
+    /// let v = IntVec::<Dynamic>::new(15);
     ///
     /// assert_eq!(15, v.bit_width());
     /// ```
@@ -125,9 +98,9 @@ impl IntVec<Dynamic> {
     /// # Examples
     ///
     /// ```
-    /// use succinct_neo::int_vec::{IntVec, IntAccess};
+    /// use succinct_neo::int_vec::{Dynamic, IntVec, IntVector};
     ///
-    /// let mut v = IntVec::new(10);
+    /// let mut v = IntVec::<Dynamic>::new(10);
     /// v.push(25);
     /// v.push(8);
     /// v.push(60);
@@ -165,154 +138,6 @@ impl IntVec<Dynamic> {
         self.size += 1;
     }
 
-    /// Grants access to the underlying slice where the bits are saved.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use succinct_neo::int_vec::{IntVec, IntAccess};
-    ///
-    /// let mut v = IntVec::new(32);
-    /// v.push(125);
-    /// v.push(1231);
-    ///
-    /// assert_eq!((1231 << 32) | 125, v.raw_data()[0]);
-    /// ```
-    #[inline]
-    pub fn raw_data(&self) -> &[usize] {
-        &self.data
-    }
-
-    #[inline]
-    /// Checks whether this has no integers saved.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use succinct_neo::int_vec::{IntVec, IntAccess};
-    ///
-    /// let mut v = IntVec::new(32);
-    ///
-    /// assert!(v.is_empty());
-    ///
-    /// v.push(125);
-    ///
-    /// assert!(!v.is_empty());
-    /// ```
-    pub fn is_empty(&self) -> bool {
-        self.size == 0
-    }
-
-    #[inline]
-    /// The amount of integers saved in this vector.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use succinct_neo::int_vec::IntVec;
-    ///
-    /// let mut v = IntVec::new(10);
-    /// for i in 0..20 {
-    ///     v.push(i);
-    /// }
-    ///
-    /// assert_eq!(20, v.len());
-    ///
-    /// ```
-    pub fn len(&self) -> usize {
-        self.size
-    }
-
-    /// An iterator over this vector's saved integers.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use succinct_neo::int_vec::IntVec;
-    ///
-    /// let mut v = IntVec::new(10);
-    /// for i in (0..20).rev() {
-    ///     v.push(i);
-    /// }
-    ///
-    /// assert!(Iterator::eq((0..20).rev(), v.iter()))
-    /// ```
-    #[inline]
-    pub fn iter(&self) -> Iter {
-        Iter { i: 0, v: self }
-    }
-
-    /// Gets an integer of the given bit width from an index.
-    ///
-    /// This gets the `width` bits at bit index `index width`.
-    ///
-    /// # Arguments
-    ///
-    /// * `index` - The index to read from.
-    /// * `width` - The bit width of integers.
-    ///
-    /// # Safety
-    ///
-    /// If `n` is the amount of bits stored in this vector, then for `index` and `width`
-    /// `(index + 1) * width < n` must hold.
-    ///
-    unsafe fn get_unchecked_with_width(&self, index: usize, width: usize) -> usize {
-        let index_block = (index * width) / Self::block_width();
-        let index_offset = (index * width) % Self::block_width();
-
-        // If we're on the border between blocks
-        if index_offset + width >= Self::block_width() {
-            let fitting_bits = Self::block_width() - index_offset;
-            let remaining_bits = width - fitting_bits;
-            let lo = self.data[index_block] >> index_offset;
-            let mask = (1 << remaining_bits) - 1;
-            let hi = self.data[index_block + 1] & mask;
-            return (hi << fitting_bits) | lo;
-        }
-
-        let mask = (1 << width) - 1;
-        (self.data[index_block] >> index_offset) & mask
-    }
-
-    /// Sets an integer of the given bit width at an index.
-    ///
-    /// This replaces the `width` bits at bit index `index width`.
-    ///
-    /// # Arguments
-    ///
-    /// * `index` - The index to write to.
-    /// * `width` - The bit width of integers.
-    ///
-    /// # Safety
-    ///
-    /// If `n` is the amount of bits stored in this vector, then for `index` and `width`
-    /// `(index + 1) * width < n` must hold.
-    /// In addition, `value` must fit into `width` bits.
-    ///
-    unsafe fn set_unchecked_with_width(&mut self, index: usize, value: usize, width: usize) {
-        let mask = (1 << width) - 1;
-        let value = value & mask;
-        let index_block = (index * width) / Self::block_width();
-        let index_offset = (index * width) % Self::block_width();
-
-        // If we're on the border between blocks
-        if index_offset + width >= Self::block_width() {
-            let fitting_bits = Self::block_width() - index_offset;
-            unsafe {
-                let lower_block = self.data.get_unchecked_mut(index_block);
-                *lower_block &= !(mask << index_offset);
-                *lower_block |= value << index_offset;
-                let higher_block = self.data.get_unchecked_mut(index_block + 1);
-                *higher_block &= !(mask >> fitting_bits);
-                *higher_block |= value >> fitting_bits;
-            }
-            return;
-        }
-
-        self.data[index_block] &= !(mask << index_offset);
-        self.data[index_block] |= value << index_offset;
-    }
-
     /// Modifies this vector to require the minimum amount of bits per saved element.
     ///
     /// This searches for the largest element in the vector. It then saves all saved all integers
@@ -321,9 +146,9 @@ impl IntVec<Dynamic> {
     /// # Examples
     ///
     /// ```
-    /// use succinct_neo::int_vec::IntVec;
+    /// use succinct_neo::int_vec::{Dynamic, IntVec};
     ///
-    /// let mut v = IntVec::with_capacity(9, 25);
+    /// let mut v = IntVec::<Dynamic>::with_capacity(9, 25);
     ///
     /// // All these numbers should take 3 bits to save
     /// for i in (0..50).step_by(2) {
@@ -354,45 +179,19 @@ impl IntVec<Dynamic> {
             }
         }
     }
-
-    /// Shrinks the allocated backing storage behind this int vector to fit the amount of saved
-    /// integers.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use succinct_neo::int_vec::IntVec;
-    ///
-    /// let mut v = IntVec::with_capacity(5, 200);
-    ///
-    /// // All these numbers should take 3 bits to save
-    /// for i in (0..50) {
-    ///     v.push(i % 8)
-    /// }
-    ///
-    /// v.shrink_to_fit();
-    ///
-    /// // 50 numbers each using 5 bits is 250 bits of storage.
-    /// // This fits into 4 * 64 bit blocks (= 256 bits), which in total would fit 51 integers.
-    /// assert_eq!(51, v.capacity());
-    /// ```
-    pub fn shrink_to_fit(&mut self) {
-        let required_blocks = Self::num_required_blocks::<usize>(self.size, self.width);
-        self.data.truncate(required_blocks);
-        self.data.shrink_to_fit();
-        self.recalculate_capacity();
-    }
 }
-impl IntAccess for IntVec<Dynamic> {
+
+impl IntVector for IntVec<Dynamic> {
     unsafe fn get_unchecked(&self, index: usize) -> usize {
         self.get_unchecked_with_width(index, self.width)
     }
 
     fn get(&self, index: usize) -> usize {
-        if index >= self.len() {
-            panic!("length is {} but index is {index}", self.len())
-        }
-
+        assert!(
+            index < self.len(),
+            "length is {} but index is {index}",
+            self.len()
+        );
         unsafe { self.get_unchecked(index) }
     }
 
@@ -401,82 +200,203 @@ impl IntAccess for IntVec<Dynamic> {
     }
 
     fn set(&mut self, index: usize, value: usize) {
-        if index >= self.len() {
-            panic!("length is {} but index is {index}", self.len())
-        }
-        if value >= (1 << self.width) {
-            panic!("value {value} too large for {}-bit integer", self.width)
-        }
+        assert!(
+            index < self.len(),
+            "length is {} but index is {index}",
+            self.len()
+        );
+        assert!(
+            value < (1 << self.width),
+            "value {value} too large for {}-bit integer",
+            self.width
+        );
         unsafe { self.set_unchecked(index, value) }
     }
-}
 
-impl IntoIterator for IntVec<Dynamic> {
-    type Item = usize;
-
-    type IntoIter = IntoIter;
-
-    fn into_iter(self) -> Self::IntoIter {
-        IntoIter { i: 0, v: self }
+    fn len(&self) -> usize {
+        self.size
     }
 }
 
-impl<'a> IntoIterator for &'a IntVec<Dynamic> {
-    type Item = usize;
+#[cfg(test)]
+mod test {
+    use crate::int_vec::{Dynamic, IntVec, IntVector};
 
-    type IntoIter = Iter<'a>;
+    #[test]
+    fn basics_test() {
+        let mut v = IntVec::<Dynamic>::new(4);
+        assert_eq!(0, v.len(), "int vec size not 0");
+        assert!(v.is_empty(), "int vec not empty");
 
-    fn into_iter(self) -> Self::IntoIter {
-        Iter { i: 0, v: self }
+        v.push(1);
+        v.push(2);
+        v.push(3);
+        v.push(4);
+
+        assert_eq!(4, v.len(), "int vec size not 4");
+        assert!(!v.is_empty(), "int vec not empty");
+
+        assert_eq!(0x4321, v.raw_data()[0], "backing data incorrect");
+        println!("{v:?}")
     }
-}
 
-pub struct IntoIter {
-    i: usize,
-    v: IntVec<Dynamic>,
-}
+    #[test]
+    fn push_test() {
+        let mut v = IntVec::<Dynamic>::new(23);
+        v.push(1);
+        v.push(2);
+        v.push(3);
+        v.push(4);
 
-impl Iterator for IntoIter {
-    type Item = usize;
+        assert_eq!(1, v.get(0));
+        assert_eq!(2, v.get(1));
+        assert_eq!(3, v.get(2));
+        assert_eq!(4, v.get(3));
+    }
 
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.i == self.v.len() {
-            return None;
+    #[test]
+    fn set_test() {
+        let mut v = IntVec::<Dynamic>::new(7);
+        for _ in 0..50 {
+            v.push(1);
         }
 
-        let res = self.v.get(self.i);
-        self.i += 1;
-        Some(res)
-    }
-}
-
-impl ExactSizeIterator for IntoIter {
-    fn len(&self) -> usize {
-        self.v.len() - self.i
-    }
-}
-
-pub struct Iter<'a> {
-    i: usize,
-    v: &'a IntVec<Dynamic>,
-}
-
-impl Iterator for Iter<'_> {
-    type Item = usize;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.i == self.v.len() {
-            return None;
+        for (expected, actual) in std::iter::repeat(1).zip(&v) {
+            assert_eq!(expected, actual)
         }
 
-        let res = self.v.get(self.i);
-        self.i += 1;
-        Some(res)
-    }
-}
+        for (i, val) in (0..50).enumerate() {
+            v.set(i, val);
+        }
 
-impl ExactSizeIterator for Iter<'_> {
-    fn len(&self) -> usize {
-        self.v.len() - self.i
+        for (expected, actual) in (0..50).zip(&v) {
+            assert_eq!(expected, actual)
+        }
+    }
+
+    #[test]
+    fn get_test() {
+        let mut v = IntVec::<Dynamic>::new(7);
+        let mut test_v = Vec::new();
+        for i in 0..30 {
+            v.push(3 * i);
+            test_v.push(3 * i);
+        }
+
+        for (i, actual) in test_v.into_iter().enumerate() {
+            assert_eq!(v.get(i), actual);
+        }
+    }
+
+    #[test]
+    fn iter_test() {
+        let mut v = IntVec::<Dynamic>::new(8);
+
+        for i in 0..20 {
+            v.push(i)
+        }
+
+        let mut iter = v.iter();
+        assert_eq!(20, iter.len(), "incorrect iterator length");
+        for (expect, actual) in (0..).zip(&mut iter) {
+            assert_eq!(expect, actual, "value at index {expect} incorrect")
+        }
+
+        assert_eq!(None, iter.next());
+    }
+
+    #[test]
+    fn into_iter_test() {
+        let mut v = IntVec::<Dynamic>::new(12);
+        let mut test_v = Vec::new();
+        let mut i = 1;
+        for _ in 0..10 {
+            v.push(i);
+            test_v.push(i);
+            i = (i << 1) | 1;
+        }
+
+        let mut iter = v.into_iter();
+        assert_eq!(10, iter.len(), "incorrect iterator length");
+        for (expect, actual) in test_v.into_iter().zip(&mut iter) {
+            assert_eq!(expect, actual);
+        }
+
+        assert_eq!(None, iter.next());
+    }
+
+    #[test]
+    #[should_panic]
+    fn get_out_of_bounds_test() {
+        let v = IntVec::<Dynamic>::new(7);
+        v.get(10);
+    }
+
+    #[test]
+    #[should_panic]
+    fn set_out_of_bounds_test() {
+        let mut v = IntVec::<Dynamic>::new(7);
+        v.set(10, 10);
+    }
+
+    #[test]
+    #[should_panic]
+    fn set_too_large_number_test() {
+        let mut v = IntVec::<Dynamic>::new(7);
+        v.push(0);
+        v.set(0, 100000000);
+    }
+
+    #[test]
+    #[should_panic]
+    fn push_too_large_number_test() {
+        let mut v = IntVec::<Dynamic>::new(7);
+        v.push(100000000);
+    }
+
+    #[test]
+    fn bit_compress_test() {
+        let mut v = IntVec::<Dynamic>::with_capacity(9, 25);
+
+        // 25 * 9 = 225, which fits into 4 64-bit numbers (= 256 bits).
+        // So the capacity should be 256 / 9 = 28
+        assert_eq!(28, v.capacity, "incorrect capacity before compression");
+
+        // All these numbers should take 3 bits to save
+        for i in (0..50).step_by(2) {
+            v.push(i % 8)
+        }
+
+        v.bit_compress();
+
+        assert_eq!(3, v.width, "incorrect word width after compression");
+
+        // We were at 256 bits before with a bit size of 3.
+        // So 256 / 3 = 85
+        assert_eq!(85, v.capacity, "incorrect capacity after compression");
+        assert_eq!(25, v.len(), "incorrect length after compression");
+
+        for i in 0..v.len() {
+            assert_eq!((2 * i) % 8, v.get(i), "incorrect value at index {i}")
+        }
+    }
+
+    #[test]
+    fn shrink_to_fit_test() {
+        let mut v = IntVec::<Dynamic>::with_capacity(9, 200);
+
+        // 200 * 9 = 1800, which fits into 29 64-bit numbers (= 1856 bits).
+        // So the capacity should be 1856 / 9 = 206
+        assert_eq!(206, v.capacity, "incorrect capacity before shrink");
+
+        for i in 0..50 {
+            v.push(i)
+        }
+
+        v.shrink_to_fit();
+
+        // We now have 50 elements in the vector, taking up 50 * 9 = 450 bits and fitting into
+        // 8 * 64 bit blocks = 512 bits. These fit 512 / 9 = 56 integers in total.
+        assert_eq!(56, v.capacity, "incorrect capacity after shrink");
     }
 }
