+++ succinct_neo/src/int_vec/traits.rs
@@ -1,6 +1,6 @@
 
 /// Allows access to integers in a datastructure
-pub trait IntAccess {
+pub trait IntVector {
     /// Gets an integer without making any checks for bounds etc.
     ///
     /// # Safety
@@ -23,5 +23,43 @@ pub trait IntAccess {
 
     /// Sets the integer at an index to the given value while checking for bounds and other requirements.
     fn set(&mut self, index: usize, value: usize);
+
+    /// The amount of integers saved in this vector.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use succinct_neo::int_vec::{Dynamic, IntVec, IntVector};
+    ///
+    /// let mut v = IntVec::<Dynamic>::new(10);
+    /// for i in 0..20 {
+    ///     v.push(i);
+    /// }
+    ///
+    /// assert_eq!(20, v.len());
+    ///
+    /// ```
+    fn len(&self) -> usize;
+
+
+    /// Checks whether this has no integers saved.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use succinct_neo::int_vec::{Dynamic, IntVec, IntVector};
+    ///
+    /// let mut v = IntVec::<Dynamic>::new(32);
+    ///
+    /// assert!(v.is_empty());
+    ///
+    /// v.push(125);
+    ///
+    /// assert!(!v.is_empty());
+    /// ```
+    #[inline]
+    fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
 }
